+++ CHANGELOG.md
@@ -1,6 +1,32 @@
 Change Log for libSplash
 ================================================================
 
+Release 1.2.0
+-------------
+**Date:** 2014-04-08
+
+**New Features**
+
+ - splash2xdmf python script to create XDMF descriptions for libSplash files
+   with domain information
+ - allow to read/write attributes at groups
+
+
+**Interface Changes**
+
+ - major interface change: simplified interfaces as most read/write routines
+   now use a Selection class for describe the source layout
+ - add a finalize() call to ParallelDataCollector to free MPI resources
+
+
+**Bug Fixes**
+
+ - fix automatic chunking algorithm
+ - only use HDF5 hyperslaps if necessary for improved performance
+ - fix bug in getMaxID and getEntryIDs for ParallelDataCollector
+
+
+
 Release 1.1.1
 -------------
 **Date:** 2014-02-20